pyproject.toml
# This file defines the workspace for the a2a-friend-scheduling project.
# It tells uv that this project is a monorepo containing multiple packages.

[project]
name = "a2a-friend-scheduling"
version = "0.1.0"
description = "A multi-agent system for scheduling pickleball games."
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
    # Shared ADK & A2A Dependencies
    "google-adk>=1.3.0",
    "a2a-sdk==0.2.5",
    "nest-asyncio==1.6.0",
    "python-dotenv",
    "click",
    "uvicorn",
    "google-generativeai",
    "httpx",
    "google-auth",
    "google-auth-oauthlib",
    "google-api-python-client",
    "beautifulsoup4",
    "httplib2",

]
